pyproject.toml
[project]
name = "my_project"
version = "0.1.0"
description = "Тестовый проект"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "numpy",
    "python-dotenv",
    "psycopg2-binary",
    "matplotlib",
    "statsmodels",
    "prophet",   
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]  
build-backend = "setuptools.build_meta" 

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "isort",
    "flake8",
]